tests: unit v integration seperation for preflight tests
