+++ pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+markers =
+    unit: no infrastructure required — runs in CI on every PR
+    integration: requires Docker/container to be running
+    load: load test subset of integration (implies integration)
+    stress: stress test subset of integration (implies integration)
